Refactored the jsx so that it follows the dispatcher pattern
